Remove templates GCGLSpan & GCGLSpanTuple from GraphicsTypesGL.h https://bugs.webkit.org/show_bug.cgi?id=250817 rdar://problem/104407319
Reviewed by Jer Noble.

To improve compile time of WebCoreArgumentCoders.h move the templates GCGLSpan &
GCGLSpanTuple into their own header to be included by implementations of
GraphicsContextGL. So GraphicsContextGL.h can compile with just forward
declarations, move query functions getFloat, getBoolean, et al, from
GraphicsContextGL.h to GraphicsContextGL.cpp

Canonical link: https://commits.webkit.org/259124@main
